backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
matplotlib==3.8.2
numpy==1.26.2
pyyaml==6.0.1
pillow==10.1.0
git+https://github.com/hanruihua/ir-sim.git
